Removed Byteflow and FlowInfo and moved respective methods to SCFG
